backend/audio_transcription/requirements.txt
fastapi
uvicorn[standard]
python-multipart
boto3
google-generativeai
python-dotenv
PyJWT